split KeplerGLBase
