+++ .gitignore
@@ -1,3 +1,7 @@
+# Manifest
+Manifest.toml
+test/Manifest.toml
+
 # token file
 token.jl
