+++ Project.toml
@@ -4,34 +4,14 @@ authors = ["Johannes Boehm <johannes.boehm@gmail.com>"]
 version = "0.0.3"
 
 [deps]
-Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
 Blink = "ad839575-38b3-5650-b840-f874b8c74a25"
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-ColorTypes = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-GeoJSON = "61d90e0f-e114-555e-ac52-39dfb47a3ef9"
-JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
+KeplerGLBase = "aea72984-0742-4639-bc0a-1d6e7e468236"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 WebIO = "0f1e0344-ec1d-5b48-a673-e5cf874b6c29"
 
 [compat]
 Aqua = "0.8"
-Base64 = "<0.0.1, 1"
 Blink = "0.12"
-CSV = "0.10"
-ColorTypes = "0.11"
-Colors = "0.12"
-DataFrames = "1.6"
-Dates = "<0.0.1, 1"
-GeoJSON = "0.7"
-JSON3 = "1.13"
-Random = "<0.0.1, 1"
-Tables = "1.10"
-UUIDs = "<0.0.1, 1"
 WebIO = "0.8"
 julia = "^1.6"
