--- assets/example_data/trip_example.geojson
@@ -1,19 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {
-        "vendor": "A"
-      },
-      "geometry": {
-        "type": "LineString",
-        "coordinates": [
-          [-74.20986, 40.81773, 0, 1564184363],
-          [-74.20987, 40.81765, 0, 1564184396],
-          [-74.20998, 40.81746, 0, 1564184409]
-        ]
-      }
-    }
-  ]
-}
